refactor: narrow Inventory.gold and TransferGold.amount from i64 to i32
WGSL has no atomic i64, so either gold narrows to i32 or we emulate i64
atomics for the transfer_gold rule on GPU. i32's ±2.1B range is well
above any practical economic-sim value, and the WGSL emitter was ALREADY
using i32 for TransferGold.amount on GPU (emit_physics_wgsl.rs:813),
creating a latent CPU/GPU type mismatch. This narrowing aligns them.

Touches:
- assets/sim/events.sim: EffectGoldTransfer.amount i64 → i32 (DSL source)
- crates/engine/src/state/agent_types.rs: Inventory.gold i64 → i32
- crates/engine/src/schema_hash.rs: both Inventory and TransferGold strings
- crates/engine/src/ability/program.rs: EffectOp::TransferGold variant
- crates/engine/tests/cast_handler_gold.rs: test helper signatures
- Regenerated: engine_generated events + engine/generated/physics/transfer_gold

Schema hash bumps — pre-1.0 so save compatibility is not a concern.

+++ crates/engine/src/ability/program.rs
@@ -14,8 +14,9 @@
 //! - The Meta effect (`CastAbility`) is what makes the subsystem recursive.
 //!   Recursion is bounded by the engine's `MAX_CASCADE_ITERATIONS = 8` —
 //!   Task 18 pins the depth budget.
-//! - `TransferGold.amount` is `i64` (signed) so debt / refunds round-trip
-//!   through the same op. `Inventory.gold` is also `i64` (state port Task H).
+//! - `TransferGold.amount` is `i32` (signed) so debt / refunds round-trip
+//!   through the same op. `Inventory.gold` is also `i32` (narrowed 2026-04-22
+//!   for GPU atomic compatibility; was i64).
 
 use crate::ability::AbilityId;
 use smallvec::SmallVec;
@@ -97,8 +98,9 @@ pub enum EffectOp {
 
     // --- World (2) ---
     /// Move gold between caster and target via `cold_inventory[slot].gold`.
-    /// Signed i64 — debt is allowed; negative `amount` reverses the flow.
-    TransferGold { amount: i64 } = 5,
+    /// Signed i32 — debt is allowed; negative `amount` reverses the flow.
+    /// Narrowed from i64 on 2026-04-22 for GPU atomic compatibility.
+    TransferGold { amount: i32 } = 5,
 
     /// Adjust symmetric `cold_standing` between caster and target by `delta`.
     /// Clamped to `[-1000, 1000]` inside the handler.
@@ -145,9 +147,10 @@ mod tests {
 
     #[test]
     fn effect_op_size_budget() {
-        // 16-byte budget: largest payload is TransferGold { amount: i64 } at
-        // 8 bytes plus a u8 discriminant, padded to 16. Anything larger means
-        // we've accidentally grown a variant — catch it at the first site.
+        // 16-byte budget: largest payload is CastAbility at ~6 bytes plus a
+        // u8 discriminant, padded to 16. Anything larger means we've
+        // accidentally grown a variant — catch it at the first site.
+        // (TransferGold was i64 pre-2026-04-22; now i32.)
         let sz = std::mem::size_of::<EffectOp>();
         assert!(sz <= 16, "EffectOp grew past 16B budget: {sz}");
     }
